refactor: restructure chat graph, improve tool execution, and add API documentation
- Move attachments and stream_callback from state to RunnableConfig in agent_node and tools_node for better separation of concerns
- Add _dispatch_ui_actions method to ToolExecutor for WebSocket-based UI action dispatching
- Fix Workflow model field reference (owner → user) in canvas_agent tools
- Add drf-spectacular for OpenAPI schema generation and interactive API documentation
- Implement retry logic with exponential backoff in OpenRouterNode for improved reliability
- Enhance SafeCodeValidator with additional dangerous attributes (__self__, __func__, __closure__, etc.)
- Convert threading-based document processing to Celery tasks (process_document_task, share_document_task)
- Add skills parameter to run_engine_worker_task for skill-aware workflow execution
- Fix JSON encoding in ExecutionLogger.average_duration_ms calculation
- Improve InputSanitizationMiddleware charset handling with fallback to UTF-8
- Add thread_id to graph state initialization and track iteration count from tool_trace
- Convert synchronous notification creation to async in send_message_stream
- Add public embed_text/embed_texts aliases to HNSWKnowledgeBase for API compatibility
- Update adrf requirement from >=0.2.0 to >=0.1.0

+++ logs/logger.py
@@ -249,7 +249,7 @@ class ExecutionLogger:
                 if status == 'completed':
                     wf.successful_executions = F('successful_executions') + 1
                 if duration_ms:
-                    wf.average_duration_ms = duration_ms if not wf.average_duration_ms else (wf.average_duration_ms * 4 + duration_ms) / 5
+                    wf.average_duration_ms = int(duration_ms if not wf.average_duration_ms else (wf.average_duration_ms * 4 + duration_ms) / 5)
                 wf.save()
             await update_workflow_stats()
         except Exception: pass
